refactor(sign-in): rename component
