+++ components/index.js
@@ -35,7 +35,7 @@ export { default as QuantitySelector } from './quantity-selector';
 export { default as QrCode } from './qr-code';
 export { default as RadioButton } from './radio-button';
 export { default as RadioButtonCard } from './radio-button-card';
-export { default as Signin } from './signin';
+export { default as Signin } from './sign-in';
 export { default as StarRating } from './star-rating';
 export { default as StepProgressBar } from './step-progress-bar';
 export { default as StoreAddress } from './store-address';
